Moved classes to chopsticks package.
