+++ src/test/scala/org/kiji/chopsticks/InputOptionsSuite.scala
@@ -17,11 +17,11 @@
  * limitations under the License.
  */
 
-package org.kiji.lang
+package org.kiji.chopsticks
 
 import org.scalatest.FunSuite
 
-import org.kiji.lang.Column.InputOptions
+import org.kiji.chopsticks.Column.InputOptions
 import org.kiji.schema.filter.RegexQualifierColumnFilter
 
 class InputOptionsSuite extends FunSuite {
